+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - Handle profile chunks in regressed endpoint. ([#527](https://github.com/getsentry/vroom/pull/527))
 - Authentication support for Kafka connection. ([#530](https://github.com/getsentry/vroom/pull/530))
 - Add support for android chunks. ([#540](https://github.com/getsentry/vroom/pull/540))
+- Use generic Chunk interface in CallTreesReadJob ([#554](https://github.com/getsentry/vroom/pull/554))
 
 **Bug Fixes**:
 
@@ -52,6 +53,7 @@
 - Annotate functions to the right thread. ([#523](https://github.com/getsentry/vroom/pull/523))
 - Should write the profiler id. ([#528](https://github.com/getsentry/vroom/pull/528))
 - Prioritize stack trace rule over app.identifier ([#529](https://github.com/getsentry/vroom/pull/529))
+- Fix android callTree timestamps and make them consistent with the sample (v2) format ([#553](https://github.com/getsentry/vroom/pull/553))
 
 **Internal**:
 
@@ -127,6 +129,10 @@
 - Remove unused chunks flamegraph endpoint ([#541](https://github.com/getsentry/vroom/pull/541))
 - Pass readjob result as pointer ([#542](https://github.com/getsentry/vroom/pull/542))
 - Bump golang.org/x/crypto from 0.24.0 to 0.31.0 ([#544](https://github.com/getsentry/vroom/pull/544))
+- Add a minDepth check for function metrics ([#550](https://github.com/getsentry/vroom/pull/550))
+- Bump x/net package to fix security issue (high severity) ([#555](https://github.com/getsentry/vroom/pull/555))
+- Enforce shorter timeout for chunks download in flamegraph generation ([#557](https://github.com/getsentry/vroom/pull/557))
+- Bump actions/create-github-app-token from 1.11.2 to 1.11.3 ([#559](https://github.com/getsentry/vroom/pull/559))
 
 ## 23.12.0